% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_connect.R
\name{pep_connect}
\alias{pep_connect}
\title{Connect to PEP PostgreSQL database}
\usage{
pep_connect()
}
\value{
A connection to the PEP PostgreSQL database
}
\description{
Connect to PEP PostgreSQL database
}
\details{
Uses the .Renviron file stored in your Documents folder to connect to the database.

Please see the ReadMe file for this package (https://github.com/StacieKozHardy/pepDataConnect) or contact Stacie Hardy for help in getting your work environment set up properly.
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Do a lot of cool stuff here
...

# When you're done, disconnect from the database
pep_disconnect(con)

}
\references{
\url{https://github.com/StacieKozHardy/pepDataConnect}
}
\seealso{
\code{\link{pep_disconnect}}
}
