% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_connect.R
\name{pep_disconnect}
\alias{pep_disconnect}
\title{Disconnect from the PEP PostgreSQL database}
\usage{
pep_disconnect(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
nothing :)
}
\description{
Disconnect from the PEP PostgreSQL database
}
\details{
Closes an existing connection to the PEP PostgreSQL database. This is important to do at the end of a working session or R script,
because there is a limit to how many open connections you can have to the database at one time.
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Do a lot of cool stuff here

# When you're done, disconnect from the database
pep_disconnect(con)

}
\seealso{
\code{\link{pep_connect}}
}
