% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_jobss.R
\name{surv_jobss.geo_4abundance_images}
\alias{surv_jobss.geo_4abundance_images}
\title{Get JoBSS data: geo_4abundance_images}
\usage{
surv_jobss.geo_4abundance_images(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
A sp object containing all records in the surv_jobss.geo_4abundance_images query.
}
\description{
Return all records from surv_jobss.geo_4abundance_images as an sp object
}
\details{
This function returns the selected records from a query that was developed a stand-alone product for the Beaufort Sea ice seal abundance estimates.
The spatial data in this query reflect the area of survey effort during the 2021 JoBSS aerial survey flights.

The data in this query relate to other JoBSS data based on the following fields:
\itemize{
  \item flight, camera_view, dt
  \item image_name
}
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Get data and do a lot of cool stuff here
data <- surv_jobss.geo_4abundance_images(con)
...

# When you're done, disconnect from the database
pep_disconnect(con)

}
