% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_jobss.R
\name{surv_jobss.geo_4abundance_seals}
\alias{surv_jobss.geo_4abundance_seals}
\title{Get JoBSS data: geo_4abundance_seals}
\usage{
surv_jobss.geo_4abundance_seals(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
A sp object containing all records in the surv_jobss.geo_4abundance_seals query.
}
\description{
Return all records from surv_jobss.geo_4abundance_seals as an sp object
}
\details{
This function returns the selected records from a query that was developed a stand-alone product for the Beaufort Sea ice seal abundance estimates.
The spatial data in this query reflect the midpoints of frames in which on-effort ice seals were detected by the IR model from images collected during the 2021 JoBSS aerial survey flights.

The data in this query relate to other JoBSS data based on the following fields:
\itemize{
  \item flight, camera_view, dt
  \item image_name
  \item detection_id
}
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Get data and do a lot of cool stuff here
data <- surv_jobss.geo_4abundance_seals(con)
...

# When you're done, disconnect from the database
pep_disconnect(con)

}
