% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_jobss.R
\name{surv_jobss.geo_images_footprint}
\alias{surv_jobss.geo_images_footprint}
\title{Get JoBSS data: geo_images_footprint}
\usage{
surv_jobss.geo_images_footprint(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
A sp object containing all records in the surv_jobss.geo_images_footprint table.
}
\description{
Return all records from surv_jobss.geo_images_footprint as an sp object
}
\details{
This function returns all records in the surv_jobss.geo_images_footprint table. This table contains all the spatial footprints for the IR, RGB and UV images
that were collected during the 2021 JoBSS aerial survey flights. These footprints were generated by the post-processing script developed by Kitware as part of the KAMERA system.

The code used to import these data into the database can be found in the
\href{https://github.com/StacieKozHardy/IceSeal_JoBSS_DataManagement}{IceSeal_JoBSS_DataManagement} repository on GitHub.

The data in this table relate to other JoBSS data based on the following fields:
\itemize{
  \item flight, camera_view, dt
  \item image_name
}
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Get data and do a lot of cool stuff here
data <- surv_jobss.geo_images_footprint(con)

# When you're done, disconnect from the database
pep_disconnect(con)

}
