% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_jobss.R
\name{surv_jobss.geo_images_meta}
\alias{surv_jobss.geo_images_meta}
\title{Get JoBSS data: geo_images_meta}
\usage{
surv_jobss.geo_images_meta(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
A sp object containing all records in the surv_jobss.geo_images_meta table.
}
\description{
Return all records from surv_jobss.geo_images_meta as an sp object
}
\details{
This function returns all records in the surv_jobss.geo_images_meta table. This table contains all the data in each of the meta.json files associated with set of IR, RGB, and UV images that were collected
during the 2021 JoBSS aerial survey flights using the KAMERA system developed by Kitware.  The data are largely an output of information from the INS. The locations are created based on the coordinates
recorded to the INS, and these locations are reflect the location of the aircraft(not the location of the image or its footprint) at the time the image was taken.
All left and right camera view locations should match that of the center camera view in this data set. The code used to import these data into the database can be found in the
\href{https://github.com/StacieKozHardy/IceSeal_JoBSS_DataManagement}{IceSeal_JoBSS_DataManagement} repository on GitHub.

The effort_reconciled field was generated and updated post-survey to reflect any necessary changes to the field-recorded effort values. The code associated with these changes can be found in the
\href{https://github.com/StacieKozHardy/IceSeal_JoBSS_DataManagement}{IceSeal_JoBSS_DataManagement} repository on GitHub.

The data in this table relate to other JoBSS data based on the following fields:
\itemize{
  \item flight, camera_view, dt
  \item image_group (for joining these data more easily to the data stored in sub_jobss.tbl_images)
}
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Get data and do a lot of cool stuff here
data <- surv_jobss.geo_images_meta(con)
...

# When you're done, disconnect from the database
pep_disconnect(con)

}
