% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_jobss.R
\name{surv_jobss.tbl_detections_original2processed_rgb}
\alias{surv_jobss.tbl_detections_original2processed_rgb}
\title{Get JoBSS data: tbl_detections_original2processed_rgb}
\usage{
surv_jobss.tbl_detections_original2processed_rgb(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
A data frame object containing all records in the surv_jobss.tbl_detections_original2processed_rgb table.
}
\description{
Return all records from surv_jobss.tbl_detections_original2processed_rgb as a data frame
}
\details{
This function returns all records in the surv_jobss.tbl_detections_original2processed_rgb table. This table contains the links between the original RGB detections (which have a different unique identifier)
to the processed RGB detections. These links are made based on the id fields in the surv_jobss.tbl_detections_original_rgb and surv_jobss.tbl_detections_processed_rgb tables via this table. The processed RGB
detections can be linked to the original RGB detections based on two methods: closest bounding box or highest score. Wou will need to choose \emph{one} of these methods for joining the data.

The code used to generate these matches can be found in the
\href{https://github.com/StacieKozHardy/IceSeal_JoBSS_DataManagement}{IceSeal_JoBSS_DataManagement} repository on GitHub.

The data in this table relate to other JoBSS data based on the following fields:
\itemize{
  \item image_name
  \item processed_id (for joining records to the id field in surv_jobss.tbl_detections_processed_rgb)
  \item original_id_distance (for joining records to the id field in surv_jobss.tbl_detections_original_rgb, where the processed RGB detection best matched based on bounding box distance)
  \item original_id_score (for joining records to the id field in surv_jobss.tbl_detections_original_rgb, where the processed RGB detection best matched based on highest detection score)
}
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Get data and do a lot of cool stuff here
data <- surv_jobss.tbl_detections_original2processed_rgb(con)

# When you're done, disconnect from the database
pep_disconnect(con)

}
\seealso{
\code{\link{surv_jobss.tbl_detections_original_rgb}}
\code{\link{surv_jobss.tbl_detections_processed_rgb}}
}
