% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_jobss.R
\name{surv_jobss.tbl_detections_original_rgb}
\alias{surv_jobss.tbl_detections_original_rgb}
\title{Get JoBSS data: tbl_detections_original_rgb}
\usage{
surv_jobss.tbl_detections_original_rgb(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
A data frame object containing all records in the surv_jobss.tbl_detections_original_rgb table.
}
\description{
Return all records from surv_jobss.tbl_detections_original_rgb as a data frame
}
\details{
This function returns all records in the surv_jobss.tbl_detections_original_rgb table. This table contains all the \emph{original} detection outputs from the ice seal ir-rgb trigger detection model
associated with the RGB images. This table reflects the original, unreviewed detections.

The detection_id field was created to have a unique identifier assigned to all detections as they related to one another across all detection tables, with the exception of the original_rgb table.
The records in the original_rgb table are different and cannot be linked to the rest of the data without using the surv_jobss.tbl_detections_original2processed_rgb table. Details about how to link
these data can be found in the documentation for the surv_jobss.tbl_detections_original2processed_rgb function.

The code used to import these data into the database can be found in the
\href{https://github.com/StacieKozHardy/IceSeal_JoBSS_DataManagement}{IceSeal_JoBSS_DataManagement} repository on GitHub.

The data in this table relate to other JoBSS data based on the following fields:
\itemize{
  \item image_name
  \item id
}
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Get data and do a lot of cool stuff here
data <- surv_jobss.tbl_detections_original_rgb(con)

# When you're done, disconnect from the database
pep_disconnect(con)

}
\seealso{
\code{\link{surv_jobss.tbl_detections_original2processed_rgb}}
}
