% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_jobss.R
\name{surv_jobss.tbl_images}
\alias{surv_jobss.tbl_images}
\title{Get JoBSS data: tbl_images}
\usage{
surv_jobss.tbl_images(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
A data frame object containing all records in the surv_jobss.tbl_images table.
}
\description{
Return all records from surv_jobss.tbl_images as a data frame
}
\details{
This function returns all records in the surv_jobss.geo_images_meta table. This table contains an inventory of all the IR, RGB, and UV images that were collected
during the 2021 JoBSS aerial survey flights using the KAMERA system developed by Kitware. The following fields supplement the list of images:
\itemize{
  \item image_type - indicates if the image is ir_image, rgb_image or uv_image
  \item image_dir - provides the network path to the image; useful for copying images or ad hoc generating image lists
  \item ir_nuc - the authoritative field for designating if the IR image at that timestamp was NUC
  \item rgb_manualreview - indicates if the RGB image at that timestamp was selected for manual review
  \item ir_nuc_rf_mlp - indicates if the IR image at that timestamp was NUC based on the random forest and a multilayer-perceptron classifier method
  \item ir_nuc_svm - indicates if the IR image at that timestamp was NUC based on the SVM machine learning algorithm method
  \item image_group - a unique identifier that clusters the IR, RGB and UV images for a particular flight, camera view and date/time together;
  was created for joining these data more easily to the data stored in sub_jobss.geo_images_meta
  \item  ir_nuc_review - indicates if the IR image was selected for a manual review of IR NUC events for validating the two approaches
}

The code used to import these data into the database and process them once they were imported can be found in the
\href{https://github.com/StacieKozHardy/IceSeal_JoBSS_DataManagement}{IceSeal_JoBSS_DataManagement} repository on GitHub.

The data in this table relate to other JoBSS data based on the following fields:
\itemize{
  \item flight, camera_view, dt
  \item image_name
  \item image_group (for joining these data more easily to the data stored in sub_jobss.geo_images_meta)
}
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Get data and do a lot of cool stuff here
data <- surv_jobss.tbl_images(con)

# When you're done, disconnect from the database
pep_disconnect(con)

}
