% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_pv_cst.R
\name{surv_pv_cst.geo_polys}
\alias{surv_pv_cst.geo_polys}
\title{Get Coastal harbor seal data: geo_polys}
\usage{
surv_pv_cst.geo_polys(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
A sp object containing all records in the surv_pv_cst.geo_polys table.
}
\description{
Return all records from surv_pv_cst.geo_polys as an sp object
}
\details{
This function returns all the survey unit polygons available to be flown during coastal harbor seal surveys.

The fields in this table are as follows:
\itemize{
  \item id - unique integer identifer for each polygon
  \item polyid - unique alpha-numeric identifier for each survey unit
  \item stockid - integer identifier for the stock in which the survey unit falls
  \item stockname - name of the stock in which the survey unit falls
  \item station - name of the tidal station for which tidal covariates are calculated
  \item distance_km - distance from the tidal station
  \item iliamna - denotes if the polyid is located at Iliamna Lake (Y)
  \item glacier_name - denotes the glacial within the polygon that is surveyed for glacial harbor seal surveys
}

The data in this table relate to other coastal harbor seal survey data based on the following fields:
\itemize{
  \item id - for joining to other tables based on poly_id
  \item polyid
  \item stockid
}
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Get data and do a lot of cool stuff here
data <- surv_pv_cst.geo_polys(con)

# When you're done, disconnect from the database
pep_disconnect(con)
}
