% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_pv_cst.R
\name{surv_pv_cst.geo_track_lines}
\alias{surv_pv_cst.geo_track_lines}
\title{Get Coastal harbor seal data: geo_track_lines}
\usage{
surv_pv_cst.geo_track_lines(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
A sp object containing all records in the surv_pv_cst.geo_track_lines table.
}
\description{
Return all records from surv_pv_cst.geo_track_lines as an sp object
}
\details{
This function returns all the available tracklines from coastal harbor seal surveys (with the tracks represented as a single line record per flight).

The fields in this table are as follows:
\itemize{
  \item id - unique integer identifer for each trackline
  \item trackid - unique identifier for each flight
}

The data in this table relate to other coastal harbor seal survey data based on the following fields:
\itemize{
  \item trackid
}
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Get data and do a lot of cool stuff here
data <- surv_pv_cst.geo_track_lines(con)

# When you're done, disconnect from the database
pep_disconnect(con)
}
