% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_pv_cst.R
\name{surv_pv_cst.geo_track_pts}
\alias{surv_pv_cst.geo_track_pts}
\title{Get Coastal harbor seal data: geo_track_pts}
\usage{
surv_pv_cst.geo_track_pts(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
A sp object containing all records in the surv_pv_cst.geo_track_pts table.
}
\description{
Return all records from surv_pv_cst.geo_track_pts as an sp object and processed to provide date/times in GMT
}
\details{
This function returns all the available tracklines from coastal harbor seal surveys (with the tracks represented all GPS points records during each flight).

The fields in this table are as follows:
\itemize{
  \item id - unique integer identifer for each record
  \item gpx_file - originating file for the GPS data
  \item trackid - unique identifier for each flight
  \item segmentid - unique identifer for each record within the flight
  \item gps_dt - date/time of the GPS location; data are presented in GMT
  \item latitude - latitude for the GPS location
  \item longitude - longitude for the GPS location
  \item altitude - altitude for the GPS location
}

The data in this table relate to other coastal harbor seal survey data based on the following fields:
\itemize{
  \item trackid
}
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Get data and do a lot of cool stuff here
data <- surv_pv_cst.geo_track_pts(con)

# When you're done, disconnect from the database
pep_disconnect(con)
}
