% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_pv_cst.R
\name{surv_pv_cst.summ_count_by_polyid}
\alias{surv_pv_cst.summ_count_by_polyid}
\title{Get Coastal harbor seal data: summ_count_by_polyid}
\usage{
surv_pv_cst.summ_count_by_polyid(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
A sp object containing all records in the surv_pv_cst.summ_count_by_polyid table.
}
\description{
Return all records from surv_pv_cst.summ_count_by_polyid as an sp object and processed to provide date/times in GMT
}
\details{
This function returns all counts for each image that has been counted in the database.

The fields in this table are as follows:
\itemize{
  \item id - unique integer identifer for each record
  \item trackid - unique identifier for each flight
  \item track_rep - the survey repetition number for the survey unit (1 if the first pass, 2 if the second pass, etc.)
  \item polyid - unique alpha-numeric identifier for each survey unit
  \item non_pup - total number of non-pup harbor seals counted on the image
  \item pup - total number of pup harbor seals counted on the image
  \item count_compromised - indicates if there were any issues with the image that resulted in the count at that location being compromised
  \item effort_type - designates the type of effort under which the survey unit was flown
  \item survey_dt - the survey date/time (in GMT) assigned to the survey unit based on href{https://github.com/StacieKozHardy/CoastalPv_DataManagement/blob/main/CoastalPv_04b_CalculateSurveyDTinEffort.R}{this code}
  \item tide_height - the tidal height based on the survey_dt and the tidal station assigned to that survey unit; calculated using href{https://github.com/StacieKozHardy/CoastalPv_DataManagement/blob/main/CoastalPv_04c_CalculateTideCovariates.R}{this code}
  \item nearest_high_dt - the date/time of the nearest high tide to the survey
  \item nearest_high_height - the tidal height of the nearest high tide to the survey
  \item nearest_low_dt - the date/time of the nearest low tide to the survey
  \item nearest_high_height - the tidal height of the nearest low tide to the survey
}

The data in this table relate to other coastal harbor seal survey data based on the following fields:
\itemize{
  \item trackid, track_rep
  \item polyid, track_rep
}
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Get data and do a lot of cool stuff here
data <- surv_pv_cst.summ_count_by_polyid(con)

# When you're done, disconnect from the database
pep_disconnect(con)
}
