% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_pv_cst.R
\name{surv_pv_cst.tbl_image_exif}
\alias{surv_pv_cst.tbl_image_exif}
\title{Get Coastal harbor seal data: tbl_image_exif}
\usage{
surv_pv_cst.tbl_image_exif(con)
}
\arguments{
\item{con}{A connection string to the PEP PostgreSQL database}
}
\value{
A sp object containing all records in the surv_pv_cst.tbl_image_exif table.
}
\description{
Return all records from surv_pv_cst.tbl_image_exif as an sp object and processed to provide look-up values (instead of codes) and date/times in GMT
}
\details{
This function returns all image information from 2004 to present.

The fields in this table are as follows:
\itemize{
  \item id - unique integer identifer for each record
  \item source_file - image name and network path
  \item exif_extract_dt - date/time when the exif data were extracted (in GMT)
  \item image_name - name of the image
  \item photog_date_id - initials and date of image photographer
  \item original_dt - date/time when the image was taken based on camera (in GMT)
  \item camera_make - camera make from image exif
  \item camera_model - camera model from image exif
  \item lens_speces - lens information from image exif
  \item exposure_time_sec - exposure time from image exif
  \item f_stop - F-stop value from image_exif
  \item focal_length - focal length from image_exif
  \item iso - ISO from image exif
  \item quality - quality from image exif
  \item white_balance - white balance from image exif
  \item sharpness - sharpness from image exif
  \item focus_mode - focus mode from image exif
  \item image_width_pix - image width from image exif
  \item image_height_pix - image height from image exif
  \item gps_dt - date/time when the image was taken based on GPS (in GMT)
  \item latitude - latitude where image was taken
  \item longitude - longitude where image was taken
  \item altitude - altitude of aircraft when image was taken
  \item gps_satellites - number of GPS satellites
  \item polyid - unique alpha-numeric identifier for each survey unit; XXX if image was not used for counting
  \item interpolated - indicates if the location data were interpolated
  \item use_for_count - indicates if the image was used for counting

}

The data in this table relate to other coastal harbor seal survey data based on the following fields:
\itemize{
  \item photog_date_id - can match to track_id, as well
  \item image_name
  \item polyid
}
}
\examples{
# Connect to PEP database
con <- pep_connect()

# Get data and do a lot of cool stuff here
data <- surv_pv_cst.tbl_image_exif(con)

# When you're done, disconnect from the database
pep_disconnect(con)
}
